fix to make safety contour accept float input

+++ src/libs/mcs-widgets/DisplaySettings.js
@@ -193,9 +193,9 @@ define([
           aisCustomLayer.displayParameters = s57CustomLayer.displayParameters;
       }));
       topic.subscribe("mcs/setSafetyContour", function(){
-        _this.input_safety.value = parseInt(arguments[0][0], 10).toString();
+        _this.input_safety.value = parseFloat(arguments[0][0], 10).toString();
         var parametersArray = s57CustomLayer.displayParameters.ECDISParameters.DynamicParameters.Parameter;
-        parametersArray[_this.findParameter(parametersArray, "SafetyContour")].value = parseInt(_this.input_safety.value, 10);
+        parametersArray[_this.findParameter(parametersArray, "SafetyContour")].value = parseFloat(_this.input_safety.value, 10);
         s57CustomLayer.refresh();
         if (aisCustomLayer)
           aisCustomLayer.displayParameters = s57CustomLayer.displayParameters;
@@ -204,7 +204,7 @@ define([
     _onApplyClick: function( /*e*/ ) {
       var parametersArray = s57CustomLayer.displayParameters.ECDISParameters.DynamicParameters.Parameter;
       parametersArray[this.findParameter(parametersArray, "ShallowContour")].value = parseInt(this.input_shallow.value, 10);
-      parametersArray[this.findParameter(parametersArray, "SafetyContour")].value = parseInt(this.input_safety.value, 10);
+      parametersArray[this.findParameter(parametersArray, "SafetyContour")].value = parseFloat(this.input_safety.value, 10);
       parametersArray[this.findParameter(parametersArray, "DeepContour")].value = parseInt(this.input_deep.value, 10);
       s57CustomLayer.refresh();
       if (aisCustomLayer)
